Update and rename HOP 1.0.1 - decode.py to HOP 1.0.1.1.py
